This PR make ML model work in parallel on few samples 
Speedup is bigger for bigger repos

Command|time before|time after
----|----|----
`python -m credsweeper --path tests/samples/ -j 16 --ml_validation` |0.9s|0.71s
`python -m credsweeper --path CredData/data/387016a6/ -j 16 --ml_validation` |4.8s|1.6s

This PR also adds value to checks in `tests/test_app.py`, as value outputted after #3